+++ examples/slider.rs
@@ -48,12 +48,12 @@ fn u8_range_slider() -> impl MakeWidget {
     let range = Dynamic::new(42..=127);
     let start = range.map_each(|range| *range.start());
     let end = range.map_each(|range| *range.end());
-    (&start, &end).for_each({
+    range.set_source((&start, &end).for_each({
         let range = range.clone();
         move |(start, end)| {
             range.set(*start..=*end);
         }
-    });
+    }));
 
     let min = Dynamic::new(u8::MIN);
     let min_text = min.linked_string();